% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistic.R
\name{descriptive_statistic}
\alias{descriptive.stats}
\alias{descriptive_statistic}
\title{descriptive statistic}
\usage{
descriptive.stats(df, na.rm = FALSE)
}
\arguments{
\item{df}{data frame to describe}
}
\value{
a matrix of descriptive statistic
}
\author{
syuoni
}

