% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistic.R
\name{descriptive_stats}
\alias{descriptive.stats}
\alias{descriptive_stats}
\title{Get the descriptive statistics for a data frame}
\usage{
descriptive.stats(df, na.rm = FALSE)
}
\arguments{
\item{df}{data frame}

\item{na.rm}{remove NA if TRUE}
}
\value{
a matrix of descriptive statistics
  \item{observations}{number of samples}
  \item{mean}{mean value}
  \item{std.dev}{standard deviation}
  \item{min}{minimun value}
  \item{max}{maximun value}
}
\description{
Get the descriptive statistics for a data frame
}

